requirements.txt
cryptography==42.0.1
Django==5.0.1
django-cors-headers==4.3.1
django-rest-knox==4.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
pytest==8.0.0
pytest-django==4.7.0
